+++ collector/src/main/java/com/bitmovin/analytics/stateMachines/PlayerStateMachine.kt
@@ -280,6 +280,7 @@ class PlayerStateMachine(
         shouldStartup: Boolean,
     ) {
         transitionState(PlayerStates.SOURCE_CHANGED, oldVideoTime, null)
+        // FIXME: reset playback qualities here!!!!
         resetSourceRelatedState()
         identicalErrorReportingLimiter.reset()
         if (shouldStartup) {